requirements.txt
# Core APIs
openai>=1.54.0
python-dotenv>=1.0.0

# Agent Framework
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.40

# Tools
tavily-python>=0.3.0
wikipedia>=1.4.0

# From Project 1 (for document search)
PyPDF2>=3.0.0
faiss-cpu>=1.7.4

# UI
streamlit>=1.30.0
